Reduce SSH interactions when taste-testing / Fix race condition
Summary: The process of checking to see if the host is already being
tested on has a race condition between checking and writing the config.
The existing code connects N times:

1. Checks for timestamp file presence
2. greps for username in config
3. greps for username in config file to see if it matches the current
   user
4. Writes the new config file out

It could be up to 6, if the legacy FB stuff for matching usernames was
being used.

Instead of using transport N times, we use it once, and pile a bunch of
sequential logic in to predicate the addition of the new config file
only if the host isn't being tested by someone else. Once the files are
written, symlinks changed, we look again to see if we "won".

Additional items:

* @serialized_config is never used as a member variable, so demoted to
  a local instance variable
* Removed single quote/apostrophe from use of serialized_config. The
  command is already run with single quotes so this was effectively
  *un*quoting the text, which is dangerous
* Removed FB legacy username lookup from who_is_testing

The new code does this in one connection using conditional evaluation (
&& )

Test Plan: Local testing, more to follow.

Reviewers:

Subscribers:

Tasks:

Tags: