taste-tester test race condition fixes (#92)
Taste-tester was originally conceived as a tool for humans. Adding automation has highlighted issues when the same host is accidentally chosen. This already gives an error, explaining that someone else is testing

> Taste-testing on _hostname_
> ERROR: User someoneelse is already testing on _hostname_
> ERROR: Taste-testing failed. HOST NOT PUT IN TESTING MODE.

The existing code uses the transport (usually SSH) multiple times which is slow, and therefore more likely to hit a race condition. Current code:

1. Checks for timestamp file presence
2. greps for username in config
3. greps for username in config file to see if it matches the current user
4. Writes the new config file out

It could be up to 6 connections, if the legacy FB stuff for matching usernames was being used.

Instead of using transport 4-6 times, we use it once, and pile a bunch of sequential logic in to predicate the addition of the new config file only if the host isn't being tested by someone else. Once the files are
written, symlinks changed, we look again to see if we "won".

Additional items:

* Removed timestamp check when reading the config file for an existing user. If the current config file mentions a user, it will definitely still be taste-tested. This check didn't help anything.
* `chef_config_path` is now written to a temporary file then renamed which makes the contents atomic. Two simultaneous writes can produce garbage configuration.
* Removed most explicit `rm -f` before `ln -s`, replacing with `ln -sf` which according to POSIX specifications also unlinks but in the same command.
* Changed single quote/apostrophe from use of `serialized_config` to double quotes. The command is already run with single quotes via SSH so this was effectively *un*quoting the text, which is more dangerous than double quoting or leaving unquoted.
* `@serialized_config` is never used as a member variable, so demoted to a local instance variable
* Removed FB legacy username lookup from `who_is_testing`.
* Move test config filename into constant.
* Use constant for preable written, and matched.
* Whole stack of rubocop config changes, and updates to make the tests pass.