use chef_zero_running? rather than port_open? to determine whether chef-zero is running
Currently we're only checking whether the state file's port param has any process listening on it. This change utilizes a new helper (chef_zero_running?) to not only check that a process is listening on the port, but specifically a chef-zero proc.

Depends on https://github.com/facebook/between-meals/pull/92