Handle non-v6 environments
Everyone has a v6 address these days - localhost. But not everyone has
a real v6 address. If that's the case, the existing code doesn't work.

There's some uglywork arounds, but we never fixed this properly because
chef-zero didn't support multiple `--host` arguments at the time.

That's not been the case for a long time, so this should work now.

FB: be sure to make sure the chef-zero in chefdk2 has this. I'm 99% sure
it does, but I don't wanna break you.

Signed-off-by: Phil Dibowitz <phil@ipom.com>