Improve transition of bundle mode with intermediate state
Bundle mode does two things:

1. It improves client download times by using a single tgz bundle file
2. It improves server speed by avoiding using knife, et al to upload cookbooks. It just creates/replaces the tgz.

The problem happens when you use taste-tester in different locations.
At Facebook we allow for taste-tester to be used during server
provisioning. The flow there is that a *different* instance of taste-tester is run on the target against itself, and passes the `-m` / `--my-hostname` option to link to the named instance. The state of bundle mode can be different between these. For the duration of testing changing the default, use the new `:compatible` value which means the server will create a tgz *and* will continue to use `knife upload` et al.

Signed-off-by: Matthew Almond <malmond@fb.com>