Ignore .swp files in tgz
These end up breaking chef client if present. This can happen if actively editing files in attributes for example.

This implementation is a stop gap until I can identify the code in knife or chef that enumerates cookbooks and role directories in a stricter way.

= Before

```
[2019-08-15T08:37:14-07:00] FATAL: Chef::Exceptions::AttributeNotFound: could not find filename for attribute .default.rb.swp in cookbook fb_init
```

= After

Taste-tested host runs chef correctly.

Signed-off-by: Matthew Almond <malmond@fb.com>