Replace mash usage with simpler .to_s coercion on key on writes
This is to fix the issue reported in #115.

It is possible to modify `state.rb` to use Mash in a way that does merge keys properly. However this is more complex than neccessary. The existing code already uses `.to_s` on reads. Given a choice between the two approaches, I am chosing to keep the code simpler and remove the unneccessary import.

Signed-off-by: Matthew Almond <malmond@fb.com>