Use chefignores logic to ignore chef files
This uses existing chefignore mechanics. This looks in both the directory specified, and in it's parent which is useful because cookbooks at Facebook are stored in more than one directory and multiple paths are configured. The chefignore is in the parent.

The practical upshot is that chefignore is honored. This includes patterns for all common backups, documentation, and things like `spec/*` which stops these being "uploaded" into the tarball.

Signed-off-by: Matthew Almond <malmond@fb.com>