Excise 'run' verb
This has not worked in a *long time*. Today it produces an error like this:

```
/opt/chefdk/embedded/lib/ruby/gems/2.4.0/gems/taste_tester-0.0.14/lib/taste_tester/commands.rb:161:in `block in runchef': undefined method `run' for #<TasteTester::Host:0x000000000137de70> (NoMethodError)
        from /opt/chefdk/embedded/lib/ruby/gems/2.4.0/gems/taste_tester-0.0.14/lib/taste_tester/commands.rb:159:in `each'
        from /opt/chefdk/embedded/lib/ruby/gems/2.4.0/gems/taste_tester-0.0.14/lib/taste_tester/commands.rb:159:in `runchef'
        from /opt/chefdk/embedded/lib/ruby/gems/2.4.0/gems/taste_tester-0.0.14/bin/taste-tester:434:in `<module:TasteTester>'
        from /opt/chefdk/embedded/lib/ruby/gems/2.4.0/gems/taste_tester-0.0.14/bin/taste-tester:32:in `<top (required)>'
        from /opt/chefdk/embedded/bin/taste-tester:23:in `load'
        from /opt/chefdk/embedded/bin/taste-tester:23:in `<main>'
```

There's two paths here:

1. Fix it.
2. Remove it.

I am chosing 2. because any form of meaningful testing involves getting an interactive shell on the target to compare the before to after state.  If you are doing this, you are in a position to run the chef-client interactively, possibly through a wrapper. I am erring on the side of removing dead code.

The name of the chef client command was never used. The help output was adjusted to use site specific values. Now we add the caveat to check local documentation for the exact invocation needed.

Signed-off-by: Matthew Almond <malmond@fb.com>