Follow symlinked directories in cookbooks and other dirs (#138)
We have several distinct trees of cookbooks, some of which share content
via symlinks. `Find.find` in Ruby does not follow symlinks.

Wrap the loop with a queue of directories to look at, starting with none
(i.e. current directory). If we find a symlinked directory, add it as a
prefix to `**/*` which still returns paths relative to the `Dir.chdir`

Use of `Dir.glob` eliminates `p` vs `relative_name` distinctions.