Allow parallel testing, take #2
Summary:
Basically a re-do of #158.
This got rolled back because @babar wrote 3 different PRs and one of them was bad.
This one wasn't. Technically it wasn't good either, as it only did `test`, not really `untest`, `run` or `keeptesting`.

This re-do fixes this.

Still the same features:
* Can parametrize the number of threads in the config, or on the command line, default to the number of cores
* Reports hosts depending whether they failed ssh or something else
* Doesn't have any extra dependency, as the thread pool is done in a very simple way

Test Plan:
Ran it on my machine. Tried test and untest, both worked.

Reviewers: nreeves, dcavalca, jaymzh

Subscribers:

Tasks:

Tags:
Signed-off-by: Olivier Raginel <babar@fb.com>